requirements.txt
Flask==2.1.0
bootstrap-flask==2.1.0
google-cloud-datastore==2.8.1
google-auth==1.31.0
requests==2.26.0